requirements.txt
crewai==0.28.0
anthropic==0.18.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
beautifulsoup4==4.12.3
